refactor: split daemon runtime into airc-daemon crate (slice 5 relocation)
Per operating doc grievance §5 (CLI/Daemon Accumulating Policy):
the needed crate split lists `airc-daemon` as a separate crate from
`airc-cli`. This PR relocates the daemon code; wiring `airc-store`
as the inbox backing is queued for slice 5b so review can focus on
one architectural change at a time.

## What moved

From `crates/airc-cli/src/` to `crates/airc-daemon/src/`:
- `daemon/handlers.rs` → `handlers.rs`
- `daemon/server.rs` → `server.rs`
- `daemon/state.rs` → `state.rs`
- `daemon/mod.rs` → folded into `lib.rs`
- `ipc/{client,mod,request,response,transport}.rs` → `ipc/`
- `peers_store.rs` → `peers_store.rs` (daemon owns peer registry state)

The CLI's `commands::run_daemon` now calls `airc_daemon::run`; the
CLI's RPC commands (`ping`, `status`, `msg`, `inbox`, `peer add/list`)
use `airc_daemon::DaemonClient` and the typed `Request`/`Response`
enums re-exported from `airc_daemon::ipc`. `peers_store` is also
re-exported so the CLI's `peer add`/`peer list` keep working.

`git mv` preserved per-file history. No code changes beyond
`use crate::daemon::*` → `use crate::*` inside the moved files and
the cli-side imports going through `airc_daemon::`.

## Test parity

- airc-cli bin: 45 → 17 (daemon-internal tests moved with the code)
- airc-daemon: new, 28 tests (the moved ones)
- airc-cli e2e: 3 → 3
- All other crates: unchanged

17 + 28 = 45 — no tests lost or skipped.

## Out of scope (queued for slice 5b)

- Wiring `airc-store` as the inbox backing (replaces the per-wire
  in-memory buffer with cursor-paged reads).
- Frame ↔ TranscriptEvent conversion design.

## Workspace impact

- Workspace members += airc-daemon
- airc-cli Cargo.toml deps += airc-daemon
- No new external dependencies; airc-daemon reuses what airc-cli
  already pulled (tokio, serde, uuid, etc.).
