+++ crates/airc-daemon/src/ipc/response.rs
