+++ crates/airc-daemon/src/state.rs
